refactor: type 폴더 이동
